requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.2
blinker==1.6.3
boto3==1.34.88
botocore==1.34.88
build==1.2.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
clarifai==9.9.2
clarifai-grpc==9.9.0
click==8.1.7
cohere==4.27
colorama==0.4.6
coloredlogs==15.0.1
contextlib2==21.6.0
dataclasses-json==0.6.1
deeplake==3.9.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
faiss-cpu==1.7.4
fastapi==0.110.1
fastavro==1.8.2
filelock==3.12.4
flatbuffers==24.3.25
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.29.0
googleapis-common-protos==1.61.0
greenlet==3.0.0
grpcio==1.59.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.17.3
humanfriendly==10.0
humbug==0.3.2
idna==3.4
importlib-metadata==6.8.0
importlib_resources==6.4.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kubernetes==29.0.0
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langsmith==0.1.49
lz4==4.3.3
manifest-ml==0.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
networkx==3.1
nlpcloud==1.1.44
numpy==1.26.1
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.13.3
openlm==0.0.5
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.1.1
pathos==0.3.2
pillow==10.2.0
posthog==3.5.0
pox==0.3.4
ppft==1.7.6.8
protobuf==4.24.4
pulsar-client==3.4.0
pyarrow==13.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
pymongo==4.7.0
PyMySQL==1.1.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-rapidjson==1.12
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.4.2
rpds-py==0.10.6
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.0
schema==0.7.5
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.22
sqlitedict==2.1.0
starlette==0.37.2
streamlit==1.33.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.1.0
tornado==6.3.3
tqdm==4.66.2
transformers==4.34.0
tritonclient==2.34.0
typer==0.12.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
uuid==1.30
uvicorn==0.29.0
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.2
zipp==3.17.0
